Added comments and suggestions
Maybe we should add in a third argument to addItem() that gives us a direct reference to the slider or toggle or whatever that will actually give us the parameter value. Also possibly switch to a GridBagLayout for tabs since right now they always fill the available space unless they overfill it.